Moving helpers & split into new file
